feat(react-x): add 'no-unnecessary-use-effect'
