+++ packages/plugins/eslint-plugin/src/configs/recommended.ts
@@ -14,7 +14,6 @@ export const rules = {
   ...dom.rules,
   ...webApi.rules,
 
-  "@eslint-react/hooks-extra/no-direct-set-state-in-use-effect": "warn",
   "@eslint-react/naming-convention/context-name": "warn",
   // "@eslint-react/naming-convention/use-state": "warn",
 } as const satisfies RulePreset;
